+++ src/router.tsx
@@ -1,5 +1,7 @@
-import { createBrowserRouter } from "react-router";
+import { createBrowserRouter, redirect } from "react-router";
 import App from "./App";
+import { AboutScreen } from "./screens/AboutScreen";
+import { BooksScreen } from "./screens/BooksScreen";
 import { ErrorScreen } from "./screens/ErrorScreen";
 
 export const router = createBrowserRouter([
@@ -7,5 +9,19 @@ export const router = createBrowserRouter([
     path: "/",
     element: <App />,
     errorElement: <ErrorScreen />,
+    children: [
+      {
+        path: "",
+        loader: () => redirect("books"),
+      },
+      {
+        path: "books",
+        element: <BooksScreen />,
+      },
+      {
+        path: "about",
+        element: <AboutScreen />,
+      },
+    ],
   },
 ]);
