+++ src/screens/AboutScreen.tsx
@@ -1,0 +1,8 @@
+export const AboutScreen = () => {
+  return (
+    <div className="about-screen">
+      <h2>About</h2>
+      <div>Made with 🧡 at Workshops.de</div>
+    </div>
+  );
+};
